// packages/ignite-adapters/src/types.d.ts
export type ReduxSliceCommandActor<SliceType extends Slice> = {
    dispatch: (event: InferStateAndEvent<SliceType>["Event"]) => void;
    getState: () => InferStateAndEvent<SliceType>["State"];
    subscribe: (listener: () => void) => () => void;
};
export type ReduxStoreCommandActor<StoreInstance extends EnhancedStore> = {
    dispatch: (event: InferStateAndEvent<StoreInstance>["Event"]) => void;
    getState: () => InferStateAndEvent<StoreInstance>["State"];
    subscribe: StoreInstance["subscribe"];
};